+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Rename class method
+- Add fetch languages method to class and example,test
+
 ## 0.1.0
 
 - Initial version, created by Stagehand